build: db types and actions

+++ src/resolvers/mutations/index.ts
@@ -1,10 +1,15 @@
+import {
+	createCustomToken,
+	createUser,
+	deleteUser,
+	updateUser,
+} from '../../firebase/index.js';
 import repositories from '../../repositories/index.js';
-import { PubSub } from 'graphql-subscriptions';
 import {
 	MutationCreateTodoArgs,
 	MutationCreateUserArgs,
 } from '../../types/graphql.js';
-
+import { PubSub } from 'graphql-subscriptions';
 const pubsub = new PubSub();
 
 const mutations = {
@@ -27,32 +32,68 @@ const mutations = {
 			password,
 			isPrivate,
 		});
-		await repositories.userRepository.save(user);
-		return user;
+		try {
+			await repositories.userRepository.save(user);
+			const googleUser = await createUser({
+				...user,
+				displayName: user.username,
+			});
+			await repositories.userRepository.update(user.id, {
+				auth_firebase_uid: googleUser.uid,
+			});
+			return createCustomToken(googleUser.uid);
+		} catch (error) {
+			console.log('error: ', error);
+			throw error;
+		}
 	},
-	updateUser: async (
-		_: any,
-		{ id, username, email, firstName, lastName, password, isPrivate }
-	) => {
+	updateUser: async (_: any, args, ctx) => {
+		const { id, username, email, firstName, lastName, password, isPrivate } =
+			args;
 		const user = await repositories.userRepository.findOneBy({ id });
 
 		if (!user) {
 			throw new Error(`User with ID ${id} not found`);
 		}
-
-		return repositories.userRepository.update(user.id, {
-			username: username ?? user.username,
-			email: email ?? user.email,
-			firstName: firstName ?? user.firstName,
-			lastName: lastName ?? user.lastName,
-			password: password ?? user.password,
-			isPrivate: isPrivate ?? user.isPrivate,
-		});
+		try {
+			await repositories.userRepository.update(user.id, {
+				username: username ?? user.username,
+				email: email ?? user.email,
+				firstName: firstName ?? user.firstName,
+				lastName: lastName ?? user.lastName,
+				password: password ?? user.password,
+				isPrivate: isPrivate ?? user.isPrivate,
+			});
+			if (user.auth_firebase_uid) {
+				await updateUser(user.auth_firebase_uid, args);
+			}
+			return user;
+		} catch (error) {
+			console.log('error: ', error);
+			throw error;
+		}
 	},
-	deleteUser: async (_: any, { id }) => {
-		await repositories.userRepository.delete(id);
+	deleteUser: async (_: any, { id }, ctx) => {
+		try {
+			const user = await repositories.userRepository.findOneBy({ id });
+			if (!user) {
+				throw new Error(`User with id: ${id} not found`);
+			}
+			await repositories.userRepository.delete(id);
+			if (user.auth_firebase_uid) {
+				await deleteUser(user.auth_firebase_uid);
+			}
+			return user;
+		} catch (error) {
+			console.log('error: ', error);
+			throw error;
+		}
 	},
-	createTodo: async (_: any, { title, userId }: MutationCreateTodoArgs) => {
+	createTodo: async (
+		_: any,
+		{ title, userId }: MutationCreateTodoArgs,
+		ctx
+	) => {
 		const user = await repositories.userRepository.findOneBy({ id: userId });
 
 		if (!user) {
@@ -71,7 +112,7 @@ const mutations = {
 
 		return todo;
 	},
-	updateTodo: async (_: any, { id, title, description, isCompleted }) => {
+	updateTodo: async (_: any, { id, title, description, isCompleted }, ctx) => {
 		const todo = await repositories.todoRepository.findOneBy({ id });
 
 		if (!todo) {
@@ -84,10 +125,15 @@ const mutations = {
 			isCompleted: isCompleted ?? todo.isCompleted,
 		});
 	},
-	deleteTodo: async (_: any, { id }) => {
+	deleteTodo: async (_: any, { id }, ctx) => {
+		const todo = await repositories.todoRepository.findOneBy({ id });
+		if (!todo) {
+			throw new Error(`Cannot find todo with id ${id}`);
+		}
 		await repositories.todoRepository.delete(id);
+		return todo;
 	},
-	createTag: async (_: any, { name, isVisible = false }) => {
+	createTag: async (_: any, { name, isVisible = false }, ctx) => {
 		const tag = repositories.tagRepository.create({
 			name,
 			isVisible,
@@ -95,7 +141,7 @@ const mutations = {
 		await repositories.tagRepository.save(tag);
 		return tag;
 	},
-	updateTag: async (_: any, { id, name, isVisible }) => {
+	updateTag: async (_: any, { id, name, isVisible }, ctx) => {
 		const tag = await repositories.tagRepository.findOneBy({ id });
 
 		if (!tag) {
@@ -107,8 +153,13 @@ const mutations = {
 			isVisible: isVisible ?? tag.isVisible,
 		});
 	},
-	deleteTag: async (_: any, { id }) => {
+	deleteTag: async (_: any, { id }, ctx) => {
+		const tag = await repositories.tagRepository.findOneBy({ id });
+		if (!tag) {
+			throw new Error(`Tag with ID ${id} not found`);
+		}
 		await repositories.tagRepository.delete(id);
+		return tag;
 	},
 };
